+++ __tests__/resources/nested-other.js
@@ -1,0 +1,9 @@
+#!/usr/bin/env node
+
+"use strict";
+
+const path = require("path");
+
+const dir = path.join(__dirname, "../other-dir");
+
+process.stdout.write(dir);